More refactoring of Cards code
